+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - **Fixes**
   - Meissonic with multiple generators  
   - Kandinsky V2.2 invalid attention processor  
+  - PixArt Sigma Small and Large loading  
 
 ## Update for 2025-06-02